backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
qdrant-client==1.6.0
sentence-transformers==2.2.2
pandas==2.1.0
numpy==1.24.3
google-generativeai==0.3.1
gunicorn==21.2.0
pytest==7.4.0
